chore: extract global user's time zone configurator
